add ucef3643 ignition modules
